fix(ci-calls): remove checkers calls dependency on CPPFLAGS
On partitioner-atf, the ci submodule cannot depend on the
ATF definition of CPPFLAGS, since some compiler flags are
passed on this variable. This will make the checkers fail
when called.
To solve this, each checker that needs the CPPFLAGS has now
an extra parameter to be passed to the function call.

Signed-off-by: Daniel Oliveira <drawnpoetry@gmail.com>