fix: erase temporary .clang_format
The .clang_format file is copied to the current working directory each
time one of the 'format' or 'format-check' rules are run, because the
clang-format tool needs this configuration file to be in the same
directory it is executed.

Previously there was a format-clean rule that would erase this file,
but it is more user-friendly to erase it every time after the commands
are executed. Otherwise, a user might neglect to call 'clean' and
mistake the file for a file in the repo and add it in a commit.

Signed-off-by: Jose Martins <josemartins90@gmail.com>