feat: add spell-checking script
Add a python script to perform spell checking.
The script supports yaml and txt files for now, but support for
additional file types is fairly simple to implement.
With yaml files a key list must be provided to restrict spell checking
to specific fields.

The script also allows for the user to specify a list of dictionaries
that should contain words that although the spell checker detects are
wrong are to be considered correct.

Example usage:
ci/spell_check.py \
    -t yaml \
    -f ./reqs/threats_and_assumptions/TA_SAF_T_004.yml \
    -k text \
    -d dictionary.txt

Signed-off-by: David Cerdeira <davidmcerdeira@gmail.com>