fix(actions/templates): init submodule in gitlint job's checkout action
The CI repo is supposed to be included as a submodule in other repos.
Therefore the template for the workflows that use the CI submodule must be
configured to initialize submodules in each job in the checkout action.
Otherwise, the job will fail due to not finding the ci submodule
artifacts.

Signed-off-by: Jose Martins <josemartins90@gmail.com>